requirements.txt
streamlit==1.38.0
langchain==0.3.0
python-dotenv==1.0.1
transformers==4.44.2
torch==2.4.1
clarifai==10.8.0